setup es6 support + node require
