+++ src/classes/Emitter.ts
@@ -1,9 +1,9 @@
 import { EventEmitter } from "events";
 import io from "socket.io-client";
-import { snowflake } from "./interfaces/snowflake";
-import Vote from "./interfaces/Vote";
-import CoolBotListConfig from "./interfaces/CoolBotListConfig";
-import { Events } from "./constants/Events";
+import { snowflake } from "../interfaces/snowflake";
+import Vote from "../interfaces/Vote";
+import CoolBotListConfig from "../interfaces/CoolBotListConfig";
+import { Events } from "../constants/Events";
 
 /**
  * The event emitter used to emit events such as votes to the user/bot.
